+++ WarehouseWebAPI/Services/ProductsServices/ProductsDeleterService.cs
@@ -12,9 +12,13 @@ namespace Services.ProductsServices
     public class ProductsDeleterService : IProductsDeleterService
     {
         private readonly IProductRepository _productRepository;
-        public ProductsDeleterService(IProductRepository productsRepository)
+        private readonly IWarehouseProductRepository _warehouseProductRepository;
+
+        public ProductsDeleterService(IProductRepository productsRepository
+            , IWarehouseProductRepository warehouseProductRepository)
         {
             _productRepository = productsRepository;
+            _warehouseProductRepository = warehouseProductRepository;
         }
         public async Task<bool> DeleteProduct(Guid? productID)
         {
@@ -28,8 +32,11 @@ namespace Services.ProductsServices
                 return false;
             }
 
-            //List<Product> products = 
-            //Chech if products with such category exist
+            List<WarehouseProduct> warehouseProducts = await _warehouseProductRepository.GetAllWarehouseProducts();
+            if (warehouseProducts.Any(warehouseProduct => warehouseProduct.ProductID == productID)) 
+            {
+                throw new ArgumentException("This Product can not be deleted since it exists in Warehouses.");
+            }
 
             await _productRepository.DeleteProduct(productID.Value);
             return true;
