+++ WarehouseWebAPI/Services/WarehousesServices/WarehousesDeleterService.cs
@@ -12,11 +12,11 @@ namespace Services.WarehousesServices
     public class WarehousesDeleterService : IWarehousesDeleterService
     {
         private readonly IWarehouseRepository _warehouseRepository;
-        //private readonly IWarehouseProductRepository _warehouseProductRepository;
-        public WarehousesDeleterService(IWarehouseRepository warehouseRepository/*, IWarehouseProductRepository warehouseProductRepository*/)
+        private readonly IWarehouseProductRepository _warehouseProductRepository;
+        public WarehousesDeleterService(IWarehouseRepository warehouseRepository, IWarehouseProductRepository warehouseProductRepository)
         {
             _warehouseRepository = warehouseRepository;
-            //_warehouseProductRepository = warehouseProductRepository;
+            _warehouseProductRepository = warehouseProductRepository;
         }
         public async Task<bool> DeleteWarehouse(Guid? warehouseID)
         {
@@ -29,11 +29,11 @@ namespace Services.WarehousesServices
             {
                 return false;
             }
-            //List<WarehouseProduct> warehouseProducts = await _warehouseProductRepository.GetAllWarehouseProducts();
-            //if (warehouseProducts.Select(warehouseProduct => warehouseProduct.WarehouseID == warehouseID).Count() != 0)
-            //{
-            //    throw new ArgumentException("This warehouse can not be deleted since there are products that belong to it.");
-            //}
+            List<WarehouseProduct> warehouseProducts = await _warehouseProductRepository.GetAllWarehouseProducts();
+            if (warehouseProducts.Select(warehouseProduct => warehouseProduct.WarehouseID == warehouseID).Count() != 0)
+            {
+                throw new ArgumentException("This warehouse can not be deleted since there are products that belong to it.");
+            }
             await _warehouseRepository.DeleteWarehouse(warehouseID.Value);
             return true;
         }
